refactor!: modularize main.py and remove locale-based sorting
BREAKING CHANGE: The --locale CLI option has been removed.

Split main.py into focused modules:
- files.py: file discovery and path utilities
- packages.py: package name validation and matching
- sorting.py: simplified C locale (ASCII) sorting

Sorting changes:
- Uses sorted() with C locale ordering to match GNU sort
- Standalone comment lines are now filtered out
- Inline comments on package lines are preserved
- Path references (./pkg, -e ./pkg) placed at end

Removes ~850 lines of locale detection, caching, and fallback logic
that was unnecessary since package names are ASCII-only (PEP 508).
