+++ pyproject.toml
@@ -113,7 +113,6 @@ force-single-line = false
 split-on-trailing-comma = true
 
 [tool.ruff.format]
-# Configure code formatting
 quote-style = "double"
 indent-style = "space"
 skip-magic-trailing-comma = false
